refactor: 클라이언트 타입 재사용
